Light refactor, moved func items into its own mod
